# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Hakeemi Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: products
    description: Product management
  - name: cart
    description: Shopping cart
  - name: orders
    description: Order management
  - name: customers
    description: Customer management
  - name: dashboard
    description: Dashboard stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers (admin only)
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics (admin only)
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Get recent orders (admin only)
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: [string, "null"]
        userId:
          type: integer
        email:
          type: string
        password:
          type: string
      required:
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: [integer, "null"]
        mustChangePassword:
          type: boolean
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - mustChangePassword
        - createdAt

    Product:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: [string, "null"]
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sku
        - name
        - nameAr
        - description
        - descriptionAr
        - price
        - category
        - imageUrl
        - inStock
        - stockQuantity
        - createdAt

    CreateProductBody:
      type: object
      properties:
        sku:
          type: [string, "null"]
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrl:
          type: string
        stockQuantity:
          type: integer
      required:
        - name
        - nameAr
        - description
        - descriptionAr
        - price
        - category
        - imageUrl

    UpdateProductBody:
      type: object
      properties:
        sku:
          type: [string, "null"]
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockQuantity:
          type: integer

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - customerEmail
        - customerPhone
        - customerAddress
        - currency
        - paymentMethod
        - items

    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - quantity
        - price

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - customerAddress
        - currency
        - paymentMethod
        - status
        - totalAmount
        - items
        - createdAt

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - price

    DashboardStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalCustomers:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
      required:
        - totalProducts
        - totalOrders
        - totalCustomers
        - totalRevenue
        - pendingOrders
        - completedOrders