# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Al-Hakeemi Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  customerAddress: string;
  currency: string;
  paymentMethod: string;
  items: OrderItemInput[];
}